+++ Runtime/Default/TestingContext.cs
@@ -6,4 +6,9 @@ namespace Moqunity
 	{
 		public TestingContext() : base(new T()) { }
 	}
+
+	public class TestingContext : Context
+	{
+		public TestingContext(Factory factory) : base(factory) { }
+	}
 }